pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-dataprep"
version = "0.1.0"
description = "Modern LLM Pre-Training Data Preparation Pipeline"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "DataPrep Team"}
]
keywords = ["llm", "pre-training", "data-preprocessing", "phi-3", "multimodal"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.20.0",
    "beautifulsoup4>=4.11.0",
    "lxml>=4.9.0",
    "pillow>=9.0.0",
    "opencv-python>=4.7.0",
    "transformers>=4.30.0",
    "sentencepiece>=0.1.99",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "nltk>=3.8",
    "langdetect>=1.0.9",
    "xxhash>=3.2.0",
    "datasketch>=1.5.9",
    "pyarrow>=12.0.0",
    "pandas>=1.5.0",
]

[project.optional-dependencies]
pdf = [
    "pypdf2>=3.0.0",
    "pdfplumber>=0.9.0",
]
latex = [
    "pylatexenc>=2.10",
]
video = [
    "av>=10.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]
quality = [
    "detoxify>=0.5.0",
    "fasttext>=0.9.2",
]
all = [
    "llm-dataprep[pdf,latex,video,quality]",
]
dev = [
    "pytest>=7.3.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/llm-dataprep"
Repository = "https://github.com/yourusername/llm-dataprep"
Documentation = "https://github.com/yourusername/llm-dataprep/blob/main/README.md"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
